Add Survey Scene Complete
1. Survey 씬 버튼 추가 (종료, 씬 전환)
2. 점수 계산 -> Firebase 연동
